解決City未發送問題

+++ src/main/java/com/example/restapi/shop/repository/ShopRepository.java
@@ -2,8 +2,9 @@ package com.example.restapi.shop.repository;
 
 import com.example.restapi.shop.entity.Shop;
 import org.springframework.data.jpa.repository.JpaRepository;
+
 import java.util.Optional;
 
-public interface ShopRepository extends JpaRepository<Shop, Long> {
+public interface ShopRepository extends JpaRepository<Shop, Integer> {
     Optional<Shop> findByAccount(String account);
 }
